Rename facade04postanalysiscleanup.py
